@@ -10,22 +10,22 @@
         "ext-ctype": "*",
         "ext-iconv": "*",
         "amphp/http-client": "^4.6",
-        "cocur/slugify": "^4.1",
-        "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^2.7",
+        "cocur/slugify": "^4.4",
+        "doctrine/annotations": "^2.0",
+        "doctrine/doctrine-bundle": "^2.10",
         "doctrine/doctrine-migrations-bundle": "^3.2",
-        "doctrine/orm": "^2.12",
+        "doctrine/orm": "^2.16",
         "friendsofsymfony/ckeditor-bundle": "^2.4",
-        "helios-ag/fm-elfinder-bundle": "^12.2",
+        "helios-ag/fm-elfinder-bundle": "^12.3",
         "phpdocumentor/reflection-docblock": "^5.3",
-        "phpstan/phpdoc-parser": "^1.6",
+        "phpstan/phpdoc-parser": "^1.23",
         "symfony/apache-pack": "^1.0",
         "symfony/asset": "6.3.*",
         "symfony/console": "6.3.*",
         "symfony/doctrine-messenger": "6.3.*",
         "symfony/dotenv": "6.3.*",
         "symfony/expression-language": "6.3.*",
-        "symfony/flex": "^2",
+        "symfony/flex": "^2.3",
         "symfony/form": "6.3.*",
         "symfony/framework-bundle": "6.3.*",
         "symfony/http-client": "6.3.*",
@@ -45,17 +45,17 @@
         "symfony/string": "6.3.*",
         "symfony/translation": "6.3.*",
         "symfony/twig-bundle": "6.3.*",
-        "symfony/ux-dropzone": "^2.2",
-        "symfony/ux-swup": "^2.9",
+        "symfony/ux-dropzone": "^2.10",
+        "symfony/ux-swup": "^2.10",
         "symfony/ux-vue": "^2.10",
         "symfony/validator": "6.3.*",
         "symfony/web-link": "6.3.*",
-        "symfony/webpack-encore-bundle": "^1.16",
+        "symfony/webpack-encore-bundle": "^2.0",
         "symfony/yaml": "6.3.*",
         "symfonycasts/reset-password-bundle": "^1.17",
         "symfonycasts/verify-email-bundle": "^1.13",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0"
+        "twig/extra-bundle": "^3.7",
+        "twig/twig": "^3.7"
     },
     "config": {
         "allow-plugins": {
@@ -112,11 +112,11 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.5",
+        "phpunit/phpunit": "^9.6",
         "symfony/browser-kit": "6.3.*",
         "symfony/css-selector": "6.3.*",
         "symfony/debug-bundle": "6.3.*",
-        "symfony/maker-bundle": "^1.0",
+        "symfony/maker-bundle": "^1.50",
         "symfony/phpunit-bridge": "^6.3",
         "symfony/stopwatch": "6.3.*",
         "symfony/web-profiler-bundle": "6.3.*"
