@@ -2,10 +2,10 @@
 
 namespace App\Controller;
 
+use App\Entity\MenuLink;
 use App\Entity\PagesList;
 use App\Services\PagesService;
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -13,27 +13,34 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class AdminPagesController extends AbstractController
 {
-    #[Route('/admin/pages', name: 'admin_pages')]
-    public function index(EntityManagerInterface $em): Response
-    {
-        $pagesRepo = $em->getRepository(PagesList::class);
-        $pages = $pagesRepo->findAll();
+    private $pageRepo;
+    private $em;
+    private $pageService;
 
+    public function __construct(EntityManagerInterface $em, PagesService $pageService) {
+        $this->em = $em;
+        $this->pageRepo = $this->em->getRepository(PagesList::class);
+        $this->pageService = $pageService;
+    }
+
+    #[Route('/admin/pages', name: 'admin_pages')]
+    public function index(): Response 
+    {
         return $this->render('pages/index.html.twig', [
-            "pages" => $pages
+            "pages" => $this->pageRepo->findAll()
         ]);
     }
 
     /* AJOUTER UNE PAGE
     ------------------------------------------------------- */
     #[Route('/admin/pages/ajouter', name: 'admin_pages_add')]
-    public function add_page(PagesService $pageService, ManagerRegistry $doctrine, Request $request) {
-        
+    public function add_page(Request $request) 
+    {
         $title = "Ajouter une page";
-        $form = $pageService->PageManager($doctrine, $request, true);
-        
+        $form = $this->pageService->PageManager($request, true);
+
         if ($form->isSubmitted() && $form->isValid()) {
-            return $this->redirectToRoute('admin_pages');
+            $this->redirectToRoute('admin_pages');
         }
 
         return $this->render('pages/page-manager.html.twig', [
@@ -45,28 +52,15 @@ class AdminPagesController extends AbstractController
     /* MODIFIER UNE PAGE
     ------------------------------------------------------- */
     #[Route('/admin/pages/modifier/{page_id}', name: 'admin_pages_modify')]
-    public function modify_page(PagesService $pageService, ManagerRegistry $doctrine, Request $request, String $page_id) {
-
+    public function modify_page(Request $request, String $page_id) 
+    {
         // Récupération du lien de la page
-        $entityManager = $doctrine->getManager();
-        $link = $entityManager->getRepository(PagesList::class)->findOneBy(['page_id' => $page_id])->getPageUrl();
-
-        // Récupération du contenu de la page
-        $pageContentFr = file_get_contents("../templates/webpages/pages/fr/" . $page_id . ".html.twig");
-        if (!$pageContentFr) {
-            $pageContentFr = '';
-            file_put_contents("../templates/webpages/pages/fr/" . $page_id . ".html.twig", $pageContentFr);
-        }
-
-        $pageContentEn = file_get_contents("../templates/webpages/pages/en/" . $page_id . ".html.twig");
-        if (!$pageContentEn) {
-            $pageContentEn = '';
-            file_put_contents("../templates/webpages/pages/en/" . $page_id . ".html.twig", $pageContentEn);
-        }
+        $page = $this->pageRepo->findOneBy(['page_id' => $page_id]);
+        $link = $page->getPageUrl();
 
         // Mise à jour du contenu
         $title = "Modifier une page";
-        $form = $pageService->PageManager($doctrine, $request, false, $page_id);   
+        $form = $this->pageService->PageManager($request, false, $page_id);   
 
         if ($form->isSubmitted() && $form->isValid()) {
             return $this->redirectToRoute('admin_pages_modify', [
@@ -77,34 +71,32 @@ class AdminPagesController extends AbstractController
         return $this->render('pages/page-manager.html.twig', [
             'title' => $title,
             'form' => $form->createView(),
-            'pageContentFr' => $pageContentFr,
-            'pageContentEn' => $pageContentEn,
             'link' => $link,
+            'name_fr' => $page->getPageName()[0],
+            'metaTitle_fr' => $page->getPageMetaTitle()[0],
+            'metaDesc_fr' => $page->getPageMetaDesc()[0],
+            'pageContent_fr' => htmlspecialchars_decode($page->getPageContent()[0]),
         ]);
     }
 
     /* SUPPRIMER UNE PAGE
     ------------------------------------------------------- */
     #[Route('/admin/pages/supprimer/{page_id}', name: 'admin_pages_delete')]
-    public function delete_page(ManagerRegistry $doctrine, string $page_id) {
-        $entityManager = $doctrine->getManager();
-        $page = $entityManager->getRepository(PagesList::class)->findOneBy(['page_id' => $page_id]);
+    public function delete_page(string $page_id) 
+    {
+        $page = $this->pageRepo->findOneBy(['page_id' => $page_id]);
+        $menuLink = $this->em->getRepository(MenuLink::class)->findBy(['page' => $page]);
 
-        if(!$page) {
+        if($page) {
+            if ($menuLink) {
+                foreach($menuLink as $link){
+                    $this->em->remove($link);
+                }
+            }
+            $this->em->remove($page);
+            $this->em->flush();
+        } else {
             throw $this->createNotFoundException("Aucune page n'a été trouvée");
-        }
-
-        $entityManager->remove($page);
-        $entityManager->flush();
-
-        // Suppression du fichier
-        $fileFr = "../templates/webpages/pages/fr/" . $page_id . ".html.twig";
-        $fileEn = "../templates/webpages/pages/en/" . $page_id . ".html.twig";
-        if (file_exists($fileFr)) {
-            unlink($fileFr);
-        }
-        if (file_exists($fileEn)) {
-            unlink($fileEn);
         }
 
         return $this->redirectToRoute('admin_pages');
