@@ -8,94 +8,145 @@ use Cocur\Slugify\Slugify;
 use DateTimeImmutable;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 
 class PostsService extends AbstractController{
 
-    function PostManager(ManagerRegistry $doctrine, Request $request, bool $newPost, string $postId = null)
-{
-    $entityManager = $doctrine->getManager();
+    private $doctrine;
+    private $posts_repo;
+    private $em;
 
-    // Initialisation du formulaire
-    if ($newPost) {
-        $post = new PostsList();
-    } else {
-        $post = $entityManager->getRepository(PostsList::class)->findOneBy(['post_id' => $postId]);
+    function __construct(ManagerRegistry $doctrine) {
+        $this->doctrine = $doctrine;
+        $this->em = $this->doctrine->getManager();
+        $this->posts_repo = $this->em->getRepository(PostsList::class);
     }
 
-    $form = $this->createForm(PostsAdminFormType::class, $post);
-    $form->handleRequest($request);
-
-    if ($form->isSubmitted() && $form->isValid()) {
-        // Récupération des données du formulaire
-        $post = $form->getData();
-
-        if (!$newPost) {
-            if (!$post) {
-                throw $this->createNotFoundException("Aucune page n'a été trouvée");
-            }
-            $postId = $post->getPostId();
-            $postFileName = $postId . ".html.twig";
-            unlink("../templates/webpages/posts/fr/" . $postFileName);
-            unlink("../templates/webpages/posts/en/" . $postFileName);
-        }
-
-        // Création du slug
+    #region Création / Modification d'un post
+    function PostManager(Request $request, Security $security, bool $newPost, string $postId = null)
+    {
         $slugify = new Slugify();
-        $slugName = $slugify->slugify($form->get('post_name')->getData());
-        $slugUrl = $slugify->slugify($form->get('post_url')->getData());
 
-        // Création de l'ID Post
-        if ($newPost) {
-            $post->setPostId($slugName);
+        // CREATION / RECUPERATION D'UN POST
+        // --------------------------------------------------------
+        $post = ($newPost) ? new PostsList() : $this->posts_repo->find($postId);
+        if (!$post) {
+            throw $this->createNotFoundException("Aucune post n'a été trouvé");
         }
 
-        // Création des dates
-        if ($newPost) {
-            $post->setCreatedAt(new DateTimeImmutable());
+        // INITIALISATION DU FORMULAIRE
+        // --------------------------------------------------------
+        $form = $this->createForm(PostsAdminFormType::class, $post);
+        $form->handleRequest($request);
+
+        // ENVOI DU FORMULAIRE
+        // --------------------------------------------------------
+        if ($form->isSubmitted() && $form->isValid()) {
+            // Récupération des données du formulaire
+            $post = $form->getData();
+
+            // Création / Modification du nom
+            $name = [$form->get('post_name_fr')->getData()];
+            $post->setPostName($name);
+
+            // Création / Modification du contenu
+            $content = [htmlspecialchars($form->get('post_content_fr')->getData())];
+            $post->setPostContent($content);
+
+            // Création / Modification du Meta Title
+            $metaTitle = [
+                !($form->get('post_meta_title_fr')->getData()) ? $name[0] : $form->get('post_meta_title_fr')->getData()
+            ];
+            $post->setPostMetaTitle($metaTitle);
+
+            // Création / Modification du Meta Desc
+            $metaDesc = [
+                $form->get('post_meta_desc_fr')->getData()
+            ];
+            $post->setPostMetaDesc($metaDesc);
+
+            // Création de l'URL
+            if ($newPost) {
+                $slugName = $slugify->slugify($name[0]);
+                if ($slugName) {
+                    $post->setPostUrl($slugName);
+                }
+            }
+
+            // Gestion des dates
+            $currentDate = new DateTimeImmutable();
+            $post->setUpdatedAt($currentDate);
+            if ($newPost) {
+                $post->setCreatedAt($currentDate);
+            }
+
+            // Création de l'auteur
+            if ($newPost) {
+                $author = $security->getUser();
+                $post->setAuthor($author);
+            }
+
+            // Création de l'image
+            $imageFile = $form->get('post_thumb')->getData();
+            if ($imageFile) {
+                $imageName = $slugify->slugify($imageFile->getClientOriginalName());
+                try {
+                    $imageFile->move(
+                        $this->getParameter('posts_img_directory'),
+                        $imageName
+                    );
+                    $post->setPostThumb($imageName);
+                } catch (\Throwable $th) {
+                    throw $th;
+                }
+            }
+            
+            // Envoi des données vers la BDD
+            $this->em->persist($post);
+            $this->em->flush();
         }
-        $post->setUpdatedAt(new DateTimeImmutable());
 
-        // Création de l'URL
-        if (!$form->get('post_url')->getData() && $newPost) {
-            $post->setPostUrl($slugName);
-        } elseif (!$form->get('post_url')->getData() && !$newPost) {
-            $post->setPostUrl($post->getPostUrl());
-        } else {
-            $post->setPostUrl($slugUrl);
-        }
-
-        // Création du Meta Title
-        $postMetaTitle = $form->get('post_meta_title')->getData();
-        if (empty($postMetaTitle)) {
-            $postMetaTitle = $newPost ? $form->get('post_name')->getData() : $post->getPostMetaTitle();
-        }
-        $post->setPostMetaTitle($postMetaTitle);
-
-        // Création / Modification de l'image principale
-        if ($form->get('post_thumb')->getData() !== null) {
-            $postImg = $form->get('post_thumb')->getData();
-            $imgFile = md5(uniqid()) . '.' . $postImg->guessExtension();
-            $postImg->move(
-                $this->getParameter('posts_img_directory'),
-                $imgFile
-            );
-            $post->setPostThumb($imgFile);
-        }
-
-        // Envoi des données vers la BDD
-        $entityManager->persist($post);
-        $entityManager->flush();
-
-        // Création du fichier TWIG
-        $file = "../templates/webpages/posts/fr/" . $post->getPostId() . '.html.twig';
-        $file_en = "../templates/webpages/posts/en/" . $post->getPostId() . '.html.twig';
-        file_put_contents($file, $form->get('post_content')->getData());
-        file_put_contents($file_en, $form->get('post_content_en')->getData());
+        return $form;
     }
+    #endregion
 
-    return $form;
-}
+    #region Affichage des posts
+    public function getAllPosts(){
+        $posts = $this->posts_repo->findAll();
+        $post_array = [];
+        foreach ($posts as $post) {
+            $last_post_array[] = [
+                'id' => $post->getId(),
+                'name' => $post->getPostName(),
+                'url' => $post->getPostUrl(),
+                'date' => $post->getCreatedAt(),
+            ];
+        };
+        return $post_array;
+    }
+    #endregion
 
+    #region Affichage des derniers posts
+    public function getLastPosts()
+    {
+        $last_posts = $this->posts_repo->findBy([], ['created_at' => 'DESC'], 3);
+        $last_post_array = [];
+        foreach ($last_posts as $post) {
+            $last_post_array[] = [
+                'id' => $post->getId(),
+                'name' => $post->getPostName(),
+                'url' => $post->getPostUrl(),
+                'date' => $post->getCreatedAt(),
+            ];
+        };
+        return $last_post_array;
+    }
+    #endregion
 
+    #region Affichage d'un post
+    public function getPost(string $post_url){
+        $post = $this->posts_repo->findOneBy(["post_url" => $post_url]);
+    }
+    #endregion
 }
